feat(pipeline): build the pipeline in acquire, activate, assemble phases
Pipeline construction was spread across YamlRouterFactory's
inner_create/inner_create_supergraph chain, with fallible resource
acquisition interleaved with service assembly and the point of no
return enforced only by comments. The new pipeline::build_pipeline
runs three phases, each under its own span: acquire gathers
everything that can fail (telemetry plugin, federation planner,
plugins, TLS/DNS client material, Redis clients, persisted-query
manifest), activate runs the plugin activation hooks that swap
process globals, and assemble builds caches and service stacks
through infallible functions. inner_create and
inner_create_supergraph are deleted; create_pipeline delegates to
build_pipeline.

One behavior fix: the APQ Redis connect ran inside RouterCreator::new,
after activation had begun, so construction could fail past the point
of no return. It now runs in acquire. Two ordering moves with no
functional effect: PersistedQueryExpander creation moved to the end of
acquire (changes only which error wins when several steps would fail),
and warmup now runs after RouterCreator::new (the pipeline-handle
gauge exists during warmup). Caches still call activate() immediately
after construction; constructing them after the meter-provider swap
binds their gauges to the provider serving this pipeline.
